requirements.txt
# Core NLP and ML Libraries
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Clustering and Dimensionality Reduction
hdbscan>=0.8.33
umap-learn>=0.5.4

# Text Processing
nltk>=3.8.1
# spacy>=3.6.0  # Optional - not currently used in codebase. Install separately if needed.
keybert>=0.7.0
langdetect>=1.0.9

# Data Collection
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-dateutil>=2.8.2

# Data Storage
sqlalchemy>=2.0.20
# Note: sqlite3 is built-in to Python, no installation needed

